Moved timepoint views etc. into a blueprint
