+++ dashboard/templates/scan/header.html
@@ -13,7 +13,7 @@
 </div>
 
 <div class="col-sm-3">
-  <a href="{{ url_for('main.timepoint', study_id=study_id, timepoint_id=scan.timepoint) }}"
+  <a href="{{ url_for('timepoints.timepoint', study_id=study_id, timepoint_id=scan.timepoint) }}"
       class="btn btn-primary">
     <span class="glyphicon glyphicon-arrow-left"></span>
     <span class="button-text">{{ scan.timepoint }}</span>
