+++ dashboard/timepoints/templates/collapsible_section.html
