+++ dashboard/timepoints/templates/incidental_finding_display.html
