+++ dashboard/timepoints/templates/sessions/empty_session.html
