+++ dashboard/timepoints/templates/sessions/scan_table.html
