+++ dashboard/users/__init__.py
@@ -1,5 +1,9 @@
 from flask import Blueprint
 
-user_bp = Blueprint('users', __name__, template_folder='templates')
+user_bp = Blueprint(
+    'users',
+    __name__,
+    template_folder='templates',
+    url_prefix='/user')
 
 from . import views
